feature/containerised-integration-test
Changes
----
* Updated: library dependencies
* Updated: Dockerfile configuration
* Added: local kind k8s cluster creation script
* Added: k8s MinIO store
* Added: k8s Spark application job
* Refactored: updated project coordinates
