+++ src/it/k8s/minio/00-namespace.yaml
@@ -1,0 +1,6 @@
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: spark-integration-test
+  labels:
+    name: spark-integration-test
