+++ src/it/k8s/minio/01-secrets.yaml
@@ -1,0 +1,9 @@
+apiVersion: v1
+kind: Secret
+metadata:
+  name: minio-basic-auth-secret
+  namespace: spark-integration-test
+type: kubernetes.io/basic-auth
+stringData:
+  username: root
+  password: root1234
