+++ src/it/k8s/minio/03-service.yaml
@@ -1,0 +1,14 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: minio-service
+  namespace: spark-integration-test
+spec:
+  type: NodePort
+  selector:
+    app: minio
+  ports:
+    - name: http
+      protocol: TCP
+      port: 9000
+      nodePort: 30090
