--- src/main/scala/org/uk/aeb/transformations/Split.scala
@@ -1,9 +1,0 @@
-package org.uk.aeb.transformations
-
-import org.apache.spark.rdd.RDD
-
-object Split {
-
-  def execute(rdd: RDD[String]): RDD[String] = rdd.flatMap(line => line.split(" "))
-
-}
