--- src/main/scala/org/uk/aeb/utils/Spark.scala
@@ -1,12 +1,0 @@
-package org.uk.aeb.utils
-
-import org.apache.spark.sql.SparkSession
-
-object Spark {
-
-  def createSparkSession(): SparkSession =
-    SparkSession
-      .builder()
-      .getOrCreate()
-
-}
