+++ src/main/scala/uk/org/aeb/transformations/Count.scala
@@ -1,0 +1,12 @@
+package uk.org.aeb.transformations
+
+import org.apache.spark.rdd.RDD
+
+object Count {
+
+  def execute(rdd: RDD[String]): RDD[(String, Long)] =
+    rdd
+      .map(word => (word, 1L))
+      .reduceByKey((a, b) => a + b)
+
+}
