+++ src/main/scala/uk/org/aeb/transformations/Split.scala
@@ -1,0 +1,9 @@
+package uk.org.aeb.transformations
+
+import org.apache.spark.rdd.RDD
+
+object Split {
+
+  def execute(rdd: RDD[String]): RDD[String] = rdd.flatMap(line => line.split(" "))
+
+}
