+++ src/main/scala/uk/org/aeb/utils/Spark.scala
@@ -1,0 +1,12 @@
+package uk.org.aeb.utils
+
+import org.apache.spark.sql.SparkSession
+
+object Spark {
+
+  def createSparkSession(): SparkSession =
+    SparkSession
+      .builder()
+      .getOrCreate()
+
+}
